moved Item interface to interfaces folder
