+++ lib/repository/interfaces/item.dart
